backend/requirements.txt
fastapi
uvicorn[standard]
exa-py 
cerebras-cloud-sdk